+++ lib/event/event.d.ts
@@ -1,2 +1,3 @@
-export declare function event<T extends Function>(target: Object, propertyKey: string | symbol, descriptor: TypedPropertyDescriptor<T>): TypedPropertyDescriptor<T>;
-export default event;
+export declare function event<T extends Function>(target: Object, propertyKey: string | symbol, descriptor?: TypedPropertyDescriptor<T>): any;
+declare const _default: PropertyDecorator;
+export default _default;