+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## v1.1
+### v1.1.0 [![10.07.2021](https://img.shields.io/date/1625941605)](https://github.com/d8corp/watch-state-decorators/tree/v1.1.0)
+- updated [watch-state](https://www.npmjs.com/package/watch-state) version
+- `event` decorator supports fields now
+- updated readme
 
 ## v1.0
 ### v1.0.2 [![14.06.2021](https://img.shields.io/date/1623680081)](https://github.com/d8corp/watch-state-decorators/tree/v1.0.2)